+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog
 * Lock: 增加expire选项,默认为30秒,用于避免PHP出错后没有解锁导致死锁
 * Db: 增加transactional方法,通过回调的方式调用事务
 * Record: 增加forUpdate,forShare,lock方法用于事务中的锁
+* Db: insertBatch增加extra参数,可用于在生成的语句后面增加"ON DUPLICATE KEY UPDATE"字句
 
 ## 0.9.22 (2016-11-22)